Refactor generator dependencies into dedicated modules
